feature(LinqProvider):SequenceEqual added

+++ Polars.CSharp/DataFrame/SeriesAccessors.cs
@@ -22,7 +22,7 @@ public partial class DataFrame : IDisposable,IEnumerable<Series>,IPolarsDataFram
     /// <summary>
     /// Get a column by its positional index (0-based).
     /// </summary>
-    public Series Column(int index)
+    public Series Column(long index)
     {
         var h = PolarsWrapper.DataFrameGetColumnAt(Handle, index);
         return new Series(h);
